Fixed middleware routing.
Now middleware matches '/' path.
